extract `process_blocks` to an io module
