+++ src/dolphin/io/__init__.py
@@ -1,5 +1,6 @@
 from ._background import *
 from ._blocks import *
 from ._core import *
+from ._process import *
 from ._readers import *
 from ._writers import *
